Refactor: move category models to category/models
